requirements.txt
feedparser
google-generativeai
streamlit
python-dotenv
arxiv